requirements.txt
torch>=1.12
numpy
matplotlib
Pillow
